+++ backend/console/views/serializers.py
@@ -5,12 +5,6 @@ from zane_api.models import Workspace
 from rest_framework import serializers
 
 
-class InstanceUserPagination(pagination.PageNumberPagination):
-    page_size = 10
-    page_size_query_param = "per_page"
-    page_query_param = "page"
-
-
 class WorkspaceListFilterSet(django_filters.FilterSet):
     name = django_filters.CharFilter(lookup_expr="icontains")
